Feature/iac iis configuration automation

+++ src/services/src/Demoulas.ProfitSharing.CachingServices/HostedServices/DuplicateNamesAndBirthdaysHostedService.cs
@@ -16,7 +16,7 @@ public sealed class DuplicateNamesAndBirthdaysHostedService : BaseCacheHostedSer
 
     protected override string BaseKeyName => "DNAB";
 
-    protected override ushort RefreshSeconds { get; set; } = ushort.MaxValue; // Every 18.2 hours refresh
+    protected override ushort RefreshSeconds { get; set; } = (ushort)short.MaxValue; // Every 9.1 hours refresh
 
     public DuplicateNamesAndBirthdaysHostedService(
         IHostEnvironment hostEnvironment,
